plugins/rdp: Fix hangs with recent FreeRDP versions
Connection to all my testing servers fails with "SERVER BUG: The support
for this feature was not announced! Use /relax-order-checks to ignore"
currently. This happens always with current FreeRDP versions after
https://github.com/FreeRDP/FreeRDP/pull/4926 has been merged. This can be
fixed by the usage of /relax-order-checks option, however, this option
should be used only if necessary needed and it should not be needed in
most of the cases. This currenlty happens always as it interfere with our
custom OrderSupports settings. Let's use the default OrderSupports
settings to fix this issue, which is possible thanks to
https://github.com/FreeRDP/FreeRDP/pull/5057.

See: https://gitlab.gnome.org/GNOME/gtk-frdp/-/issues/27